Refactor User class and add HookedUser class with asymmetric visibility properties; update tests to validate property hooks
